==> Diplom/notebooks/00_Input_data_analyse.ipynb <==
import pandas as pd
from tqdm import tqdm_notebook

train_data = pd.read_csv("../data/external/train_data.csv", sep=';')
train_labels = pd.read_csv("../data/external/train_labels.csv", sep=';')
test_data = pd.read_csv("../data/external/test_data.csv", sep=';')

# pd.set_option('max_colwidth', 8)
# pd.set_option('display.width', 500)

# параметры процедур из истории участия
train_data.head()

train_data.info()

# актуальные процедуры, которые нужно рекомендовать
test_data.head()

test_data.info()

- pn_lot_anon – анонимизированный номер связки процедура-лот;
- fz - федеральный закон, к которому относится процедура;
- region_code – код региона (справочник регионов приложен);
- min_publish_date – дата первой публикации извещения;
- purchase_name – название закупки;
- lot_name – название лота;
- lot_price – цена лота (в руб.);
- okpd2_code – код ОКПД2;
- okpd2_names – название кода ОКПД2 (разделитель “||”);
- additional_code – добавочный код (если нет кода ОКПД2, используется КТРУ для 44-ФЗ или ОКВЭД2 для 223-ФЗ);
- additional_code_names – название добавочного кода (разделитель “||”);
- item_descriptions – описание товаров (разделитель “||”).

Код ОКПД2 имеет несколько уровней вложенности. Например, «24.20.31.000 – трубы сварные для нефте- и газопроводов, наружным диаметром не более 406,4 мм, стальные». В поле okpd2_names название кода ОКПД2 всегда соответствует реальному уровню вложенности, а в okpd2_code код. ОКПД2 верхнеуровневый. В нашем примере okpd2_names сохраняется, но okpd2_code будет 24.2 (не больше одного знака после первой точки). Аналогично для additional_code и additional_code_names. 

# история участия
train_labels

train_labels.info()

- pn_lot_anon – анонимизированный номер связки процедура-лот из train_data;
- participant_inn_kpp_anon – анонимизированный ИНН_КПП поставщика;
- is_winner – 1 - если поставщик победил в соответствующей процедуре, 0 – поставщик участвовал в соответствующей процедуре;
- fz – федеральный закон, к которому относится процедура.
