==> Diplom/notebooks/01_Data_description_preprocessing.ipynb <==
import pandas as pd
from tqdm import tqdm_notebook
import warnings
warnings.filterwarnings('ignore')
import matplotlib.pyplot as plt

import re
from nltk.corpus import stopwords
from pymorphy2 import MorphAnalyzer
pymorphy2_analyzer = MorphAnalyzer()
%matplotlib inline

train_data = pd.read_csv("../data/external/train_data.csv", sep=';')
train_labels = pd.read_csv("../data/external/train_labels.csv", sep=';')
test_data = pd.read_csv("../data/external/test_data.csv", sep=';')

#Заполнение пропусков строкой 'None'
train_data = train_data.fillna('None')
test_data = test_data.fillna('None')

train_data['okpd2_or_additional_code'] = train_data[['okpd2_code', 'additional_code']].apply(lambda x: x[0] if x[1] == 'None' else x[1], axis=1)
test_data['okpd2_or_additional_code'] = test_data[['okpd2_code', 'additional_code']].apply(lambda x: x[0] if x[1] == 'None' else x[1], axis=1)

# Формирование описания тендера
train_data['text_description_tender'] = train_data['purchase_name'] + " " + train_data['lot_name'] + " " + train_data['okpd2_names'] + " " + train_data['additional_code_names'] + " " + train_data['item_descriptions']
test_data['text_description_tender'] = test_data['purchase_name'] + " " + test_data['lot_name'] + " " + test_data['okpd2_names'] + " " + test_data['additional_code_names'] + " " + test_data['item_descriptions']

train_data.shape

test_data.shape

# train_data = train_data[:10000]
# test_data = test_data[:10000]

mystopwords = stopwords.words('russian')
ru_words = re.compile("[А-Яа-я]+")

# функция фильтрующая текст от остальных символов
def words_only(text):
    return " ".join(ru_words.findall(text))

# функция лемматизации текста
# def lemmatize(text, mystem=m):
#     try:
#         return "".join(m.lemmatize(text)).strip()  
#     except:
#         return " "
    
def lemmatize(text):
    try:
        return " ".join(pymorphy2_analyzer.parse(word)[0].normal_form for word in text.split())
    except:
        return " "

# функция для удаления стоп-слов
def remove_stopwords(text, mystopwords = mystopwords):
    try:
        return " ".join([token for token in text.split() if not token in mystopwords])
    except:
        return ""

# функция для удаления дубликатов (для случая когда в text_description_tender попал одинаковый текст при объединении текста из ячеек)
def remove_word_duplicates(text):
    text_ = []
    for word in text.split():
        if word not in text_:
            text_.append(word)
    return ' '.join(text_)    

# итоговая функция обработки текста
def preprocess(text):
    return remove_word_duplicates(remove_stopwords(lemmatize(words_only(text.lower()))))

digits = re.compile("[0-9]+")
def digits_only(text):
    return int(''.join(digits.findall(text)))

%%time
# обработка текста
train_data['text_description_tender_stemmed'] = train_data.text_description_tender.apply(preprocess)
test_data['text_description_tender_stemmed'] = test_data.text_description_tender.apply(preprocess)

%%time
# приведение pn_lot_anon к числовому виду, чтобы использовать в качестве уникального идентификатора
train_data['pn_lot_id'] = train_data.pn_lot_anon.apply(digits_only)
test_data['pn_lot_id'] = test_data.pn_lot_anon.apply(digits_only)

# создаем датасеты с описаниями тендеров и сохраняем в csv файлы для дальнейшего использования
train_data_desc_stemmed = train_data[['pn_lot_id', 'text_description_tender_stemmed']]
test_data_desc_stemmed = test_data[['pn_lot_id', 'text_description_tender_stemmed']]

train_data_desc_stemmed

test_data_desc_stemmed

train_data_desc_stemmed.to_csv('../data/intermid/train_data_desc_stemmed.csv', index=False, encoding='utf-8')
test_data_desc_stemmed.to_csv('../data/intermid/test_data_desc_stemmed.csv', index=False, encoding='utf-8')
