==> Diplom/notebooks/02_svd_vectors.ipynb <==
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import TruncatedSVD

# считываем датасеты с нормированными описаниями тендеров
train_data_desc_stemmed = pd.read_csv('../data/intermid/train_data_desc_stemmed.csv')
test_data_desc_stemmed = pd.read_csv('../data/intermid/test_data_desc_stemmed.csv')

# соединяем датасеты
data = pd.concat([train_data_desc_stemmed, test_data_desc_stemmed], ignore_index=True)

print(data.shape)
data.head()

%%time
vectors = TfidfVectorizer().fit_transform(data.text_description_tender_stemmed) # строим тф-идф матрицу документ-слово
X_reduced = TruncatedSVD(n_components=100, random_state=0).fit_transform(vectors) # оставляем n главных компонент

%%time
vectors_1 = TfidfVectorizer(ngram_range=(2, 2)).fit_transform(data.text_description_tender_stemmed) # строим тф-идф матрицу документ-слово
X_reduced_1 = TruncatedSVD(n_components=100, random_state=0).fit_transform(vectors_1) # оставляем n главных компонент

# создаем датасет из идентификаторов и векторов и сохраняем в csv файл для дальнейшего использования
svd_vectors_redused  = pd.DataFrame(data['pn_lot_id']).join(pd.DataFrame(X_reduced))
svd_vectors_redused_1  = pd.DataFrame(data['pn_lot_id']).join(pd.DataFrame(X_reduced_1))

svd_vectors_redused

svd_vectors_redused_1

svd_vectors_redused.to_csv('../data/intermid/svd_vectors_redused.csv', index=False, encoding='utf-8')
svd_vectors_redused_1.to_csv('../data/intermid/svd_vectors_ngrams_redused.csv', index=False, encoding='utf-8')

# svd_vectors_redused
