==> Diplom/notebooks/04_Bert.ipynb <==
Этот ноутбук в google colab

https://colab.research.google.com/drive/1IR6_9CRrI6nUwwYxsimv-PNhy6aVD9vY?usp=sharing

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.decomposition import TruncatedSVD
from tqdm import notebook
import torch
import warnings
# import transformers
# from tensorflow.keras.layers import Input, Dense, Concatenate, Dropout, Add, dot, LSTM
warnings.filterwarnings("ignore")

import transformers

# pip install transformers
# pip install google

# pip install google

https://github.com/google-research/bert

https://huggingface.co/transformers/model_doc/bert.html#bertmodel

# считываем датасеты с нормированными описаниями тендеров

train_data_desc_stemmed = pd.read_csv('../data/intermid/train_data_desc_stemmed.csv')
test_data_desc_stemmed = pd.read_csv('../data/intermid/test_data_desc_stemmed.csv')

# соединяем датасеты

data = pd.concat([train_data_desc_stemmed, test_data_desc_stemmed], ignore_index=True)

data = data[:1000]

# инициализация модели
tokenizer = transformers.BertTokenizer(vocab_file='../download_models/uncased_L-12_H-768_A-12/vocab.txt')
config = transformers.BertConfig.from_json_file('../download_models/uncased_L-12_H-768_A-12/bert_config.json')
model = transformers.BertModel.from_pretrained('bert-base-uncased')

# tokenized = data['text_description_tender_stemmed'].apply(lambda x: tokenizer.encode(x, max_length=500, add_special_tokens = True))
tokenized = []
for str in tqdm_notebook(data['text_description_tender_stemmed']):
  tokenized.append(tokenizer.encode(str, max_length=100, add_special_tokens = True))

print("Tokenized head:", pd.DataFrame(tokenized).head())
n = max(map(len, tokenized))  # применим padding (уравняем длины исходных описаний)
for i in range(len(tokenized)):   
    tokenized[i] = tokenized[i] + [0]*(n-len(tokenized[i]))
tokenized = np.stack(tokenized)
attention_mask = np.where(tokenized != 0, 1, 0) # создадим маску (укажем нулевые и ненулевые значения)
print("Attention mask:", attention_mask.shape)

# сделаем цикл по батчам:
batch_size = 100
embeddings = []
for i in notebook.tqdm(range(tokenized.shape[0] // batch_size)):       
    batch = torch.LongTensor(tokenized[batch_size*i:batch_size*(i+1)])  # преобразуем данные в формат тензоров
    attention_mask_batch = torch.LongTensor(attention_mask[batch_size*i:batch_size*(i+1)])  # преобразуем маску
    with torch.no_grad(): # градиенты не нужны, модель BERT обучать не будем
        batch_embeddings = model(batch, attention_mask=attention_mask_batch)
        # получаем эмбеддинги для батча, передав модели данные и маску
    embeddings.append(batch_embeddings[0][:,0,:].numpy())  
        # извлекаем нужные элементы из тензора и добавляем в список эмбеддингов  

embeddings

vectors = np.concatenate(embeddings)
print(vectors.shape)

X_reduced = TruncatedSVD(n_components=100, random_state=0).fit_transform(vectors) # оставляем n главных компонент
bert_vectors_redused  = pd.DataFrame(data['pn_lot_id']).join(pd.DataFrame(X_reduced))

bert_vectors_redused

bert_vectors_redused.to_csv('../data/intermid/bert_vectors_redused.csv', index=False, encoding='utf-8')
