==> Diplom/notebooks/05_feature_engineering_p1_(ohe_fz_region_okpd2).ipynb <==
import pandas as pd
from tqdm import tqdm_notebook
import warnings
warnings.filterwarnings('ignore')
import matplotlib.pyplot as plt
import re

# считываем данные
train_data = pd.read_csv("../data/external/train_data.csv", sep=';')
train_labels = pd.read_csv("../data/external/train_labels.csv", sep=';')
test_data = pd.read_csv("../data/external/test_data.csv", sep=';')

#Заполнение пропусков строкой 'None'
train_data = train_data.fillna('None')
test_data = test_data.fillna('None')

train_data['okpd2_or_additional_code'] = train_data[['okpd2_code', 'additional_code']].apply(lambda x: x[0] if x[1] == 'None' else x[1], axis=1)
test_data['okpd2_or_additional_code'] = test_data[['okpd2_code', 'additional_code']].apply(lambda x: x[0] if x[1] == 'None' else x[1], axis=1)

digits = re.compile("[0-9]+")
def digits_only(text):
    return int(''.join(digits.findall(text)))

%%time
# приведение pn_lot_anon к числовому виду, чтобы использовать в качестве уникального идентификатора
train_data['pn_lot_id'] = train_data.pn_lot_anon.apply(digits_only)
test_data['pn_lot_id'] = test_data.pn_lot_anon.apply(digits_only)

# соединяем датасеты
data = pd.concat([train_data, test_data], ignore_index=True)
# data

from sklearn.preprocessing import OneHotEncoder
ohe = OneHotEncoder()

# преобразовываем категориальный признак фз с помощью OneHotEncoding и сохраняем в csv файл для дальнейшего использования
ohe_fz = ohe.fit_transform(data[['fz']]).toarray()
ohe_fz = pd.DataFrame(data['pn_lot_id']).join(pd.DataFrame(ohe_fz, columns=ohe.categories_[0]))
ohe_fz

ohe_fz.to_csv('../data/intermid/ohe_fz.csv', index=False, encoding='utf-8')

# преобразовываем категориальный признак региона с помощью OneHotEncoding и сохраняем в csv файл для дальнейшего использования
ohe_region = ohe.fit_transform(data[['region_code']]).toarray()
ohe_region = pd.DataFrame(data['pn_lot_id']).join(pd.DataFrame(ohe_region, columns=['region_'+item for item in ohe.categories_[0].astype(str)]))
ohe_region

ohe_region.to_csv('../data/intermid/ohe_region.csv', index=False, encoding='utf-8')

# преобразуем тип данных столбца 'okpd2_or_additional_code' в строку
data['okpd2_or_additional_code'] = data['okpd2_or_additional_code'].astype(str)

# преобразовываем категориальный признак ОКПД2 или дополнительный код с помощью OneHotEncoding и сохраняем в csv файл для дальнейшего использования
ohe_okpd2 = ohe.fit_transform(data[['okpd2_or_additional_code']]).toarray()
ohe_okpd2 = pd.DataFrame(data['pn_lot_id']).join(pd.DataFrame(ohe_okpd2, columns='code_'+ohe.categories_[0]))
ohe_okpd2

ohe_okpd2.to_csv('../data/intermid/ohe_okpd2.csv', index=False, encoding='utf-8')
