==> Diplom/notebooks/07_metrics.ipynb <==
import pandas as pd

# загружаем необходимые файлы

# recommendations = pd.read_csv("../data/processed/recommendations_svd.csv", sep=';')
# recommendations = pd.read_csv("../data/processed/recommendations_svd_v2.csv", sep=';')
# recommendations = pd.read_csv("../data/processed/recommendations_svd_v3.csv", sep=';')
# recommendations = pd.read_csv("../data/processed/recommendations_svd_v4.csv", sep=';')
# recommendations = pd.read_csv("../data/processed/recommendations_svd_v5.csv", sep=';')
# recommendations = pd.read_csv("../data/processed/recommendations_doc2vec.csv", sep=';')
# recommendations = pd.read_csv("../data/processed/recommendations_svd_ngrams.csv", sep=';')
recommendations = pd.read_csv("../data/processed/recommendations_bert.csv", sep=';')
test_labels = pd.read_csv("../data/external/test_labels.csv", sep=';')

# формируем из данных множества

true = set((test_labels['pn_lot_anon'] + "_" + test_labels['participant_inn_kpp_anon']).values)
pred = set((recommendations['actual_recommended_pn_lot'] + "_" + recommendations['inn_kpp']).values)

# количество совпадений

intersection = len(true.intersection(pred))
intersection

# рассчет метрик

print(f"Точность: {intersection / len(pred) * 100}")
print(f"Полнота: {intersection / len(true) * 100}")
