==> Diplom/notebooks/08_Visulisation.ipynb <==
import numpy as np
import pandas as pd
import re
from tqdm import tqdm_notebook
from tqdm import tqdm

pd.set_option('max_colwidth', 500)
# pd.set_option('display.width', 120)
# pd.set_option('display.max_rows', 36)

# считываем данные

train_labels = pd.read_csv("../data/external/train_labels.csv", sep=';')
train_data = pd.read_csv("../data/external/train_data.csv", sep=';')
test_data = pd.read_csv("../data/external/test_data.csv", sep=';')


recommendations_svd = pd.read_csv("../data/processed/recommendations_svd.csv", sep=';')
recommendations_doc2vec = pd.read_csv("../data/processed/recommendations_doc2vec.csv", sep=';')

# заполняем пропуски, объединяем окпд2 и дополнительный код, объединяем описания

train_data = train_data.fillna('None')
test_data = test_data.fillna('None')

train_data['okpd2_or_additional_code'] = train_data[['okpd2_code', 'additional_code']].apply(lambda x: x[0] if x[1] == 'None' else x[1], axis=1)
test_data['okpd2_or_additional_code'] = test_data[['okpd2_code', 'additional_code']].apply(lambda x: x[0] if x[1] == 'None' else x[1], axis=1)

train_data['text_description_tender'] = train_data['purchase_name'] + " " + train_data['lot_name'] + " " + train_data['okpd2_names'] + " " + train_data['additional_code_names'] + " " + train_data['item_descriptions']
test_data['text_description_tender'] = test_data['purchase_name'] + " " + test_data['lot_name'] + " " + test_data['okpd2_names'] + " " + test_data['additional_code_names'] + " " + test_data['item_descriptions']

def show_lots(user_id):
    '''
    
    функция для показа истории и рекомендаций для участника
    
    '''
    history_ = train_labels[train_labels['participant_inn_kpp_anon'] == user_id].merge(train_data, how='left', left_on='pn_lot_anon', right_on='pn_lot_anon')[['fz_x', 'region_code', 'okpd2_or_additional_code', 'text_description_tender']]
    recomended_lots_1 = recommendations_svd[recommendations_svd.inn_kpp == user_id].merge(test_data,  how='left', left_on='actual_recommended_pn_lot', right_on='pn_lot_anon')[['fz', 'region_code', 'okpd2_or_additional_code', 'text_description_tender']]
    recomended_lots_2 = recommendations_doc2vec[recommendations_doc2vec.inn_kpp == user_id].merge(test_data,  how='left', left_on='actual_recommended_pn_lot', right_on='pn_lot_anon')[['fz', 'region_code', 'okpd2_or_additional_code', 'text_description_tender']]
    history_.to_csv("../data/processed/history.csv", index=False)
    recomended_lots_1.to_csv("../data/processed/rec_1.csv", index=False)
    recomended_lots_2.to_csv("../data/processed/rec_2.csv", index=False)

    return 'Done'

# рекомендации для случайного участника

show_lots(recommendations_svd.inn_kpp.sample().values[0])

pd.read_csv("../data/processed/history.csv")

pd.read_csv("../data/processed/rec_1.csv")

pd.read_csv("../data/processed/rec_2.csv")
